Moving the test case "InvokeItemsDialog_EditValueInvoked_NoException' to UI Integration Tests
