+++ src/System.Windows.Forms.Design/tests/UnitTests/System/Windows/Forms/Design/ListControlUnboundActionListTests.cs
@@ -39,28 +39,4 @@ public sealed class ListControlUnboundActionListTests : IDisposable
         methodItem.Category.Should().Be(SR.ItemsCategoryName);
         methodItem.Description.Should().Be(SR.ListControlUnboundActionListEditItemsDescription);
     }
-
-    [WinFormsFact]
-    public void InvokeItemsDialog_EditValueInvoked_NoException()
-    {
-        using ListView listView = new();
-        using ListViewDesigner listViewDesigner = new();
-        listViewDesigner.Initialize(listView);
-        var actionList = new ListControlUnboundActionList(listViewDesigner);
-        var task = Task.Run(async () =>
-        {
-            await Task.Delay(1000).ConfigureAwait(false);
-            foreach (Form form in Application.OpenForms)
-            {
-                if (form.Text.Contains("ListViewItem"))
-                {
-                    await form.InvokeAsync(new Action(() => form.Dispose())).ConfigureAwait(false);
-                    break;
-                }
-            }
-        });
-
-        Action action = actionList.InvokeItemsDialog;
-        action.Should().NotThrow();
-    }
 }
